requirements.txt
python-dotenv==0.17.0
flask==1.1.2
pymongo
numpy==1.20.1

